Added a True Vision field
Added a circle around the enemy that will be "True Vision" if they have already spotted the player; True Vision meaning vision through walls and everything.
A function that is called every 0.3 second after the player has entered the True Vision field will check if the player still is inside the field.